+++ packages/react/src/components/Comments/FloatingComposer.tsx
@@ -3,13 +3,18 @@ import {
   DefaultBlockSchema,
   DefaultInlineContentSchema,
   DefaultStyleSchema,
+  Dictionary,
   InlineContentSchema,
   mergeCSSClasses,
   StyleSchema,
 } from "@blocknote/core";
 import { CommentsExtension } from "@blocknote/core/comments";
+import { memo, useCallback } from "react";
 
-import { useComponentsContext } from "../../editor/ComponentsContext.js";
+import {
+  Components,
+  useComponentsContext,
+} from "../../editor/ComponentsContext.js";
 import { useCreateBlockNote } from "../../hooks/useCreateBlockNote.js";
 import { useExtension } from "../../hooks/useExtension.js";
 import { useDictionary } from "../../i18n/dictionary.js";
@@ -17,6 +22,33 @@ import { CommentEditor } from "./CommentEditor.js";
 import { defaultCommentEditorSchema } from "./defaultCommentEditorSchema.js";
 import { useBlockNoteEditor } from "../../hooks/useBlockNoteEditor.js";
 import { TextSelection } from "@tiptap/pm/state";
+
+type FloatingComposerActionsProps = {
+  isFocused: boolean;
+  isEmpty: boolean;
+  onSave: () => Promise<void>;
+  Components: Components;
+  dict: Dictionary;
+};
+
+const FloatingComposerActionsComponent = memo(
+  ({ isEmpty, onSave, Components, dict }: FloatingComposerActionsProps) => (
+    <Components.Generic.Toolbar.Root
+      className={mergeCSSClasses("bn-action-toolbar", "bn-comment-actions")}
+      variant="action-toolbar"
+    >
+      <Components.Generic.Toolbar.Button
+        className={"bn-button"}
+        mainTooltip={dict.comments.save_button_text}
+        variant="compact"
+        isDisabled={isEmpty}
+        onClick={onSave}
+      >
+        {dict.comments.save_button_text}
+      </Components.Generic.Toolbar.Button>
+    </Components.Generic.Toolbar.Root>
+  ),
+);
 
 /**
  * The FloatingComposer component displays a comment editor "floating" card.
@@ -46,44 +78,34 @@ export function FloatingComposer<
     schema: comments.commentEditorSchema || defaultCommentEditorSchema,
   });
 
+  const onSave = useCallback(async () => {
+    // (later) For REST API, we should implement a loading state and error state
+    await comments.createThread({
+      initialComment: {
+        body: newCommentEditor.document,
+      },
+    });
+    comments.stopPendingComment();
+    editor.transact((tr) => {
+      tr.setSelection(TextSelection.create(tr.doc, tr.selection.to));
+    });
+    editor.focus();
+  }, [comments, newCommentEditor, editor]);
+
   return (
     <Components.Comments.Card className={"bn-thread"}>
       <CommentEditor
         autoFocus={true}
         editable={true}
         editor={newCommentEditor}
-        actions={({ isEmpty }) => (
-          <Components.Generic.Toolbar.Root
-            className={mergeCSSClasses(
-              "bn-action-toolbar",
-              "bn-comment-actions",
-            )}
-            variant="action-toolbar"
-          >
-            <Components.Generic.Toolbar.Button
-              className={"bn-button"}
-              mainTooltip={dict.comments.save_button_text}
-              variant="compact"
-              isDisabled={isEmpty}
-              onClick={async () => {
-                // (later) For REST API, we should implement a loading state and error state
-                await comments.createThread({
-                  initialComment: {
-                    body: newCommentEditor.document,
-                  },
-                });
-                comments.stopPendingComment();
-                editor.transact((tr) => {
-                  tr.setSelection(
-                    TextSelection.create(tr.doc, tr.selection.to),
-                  );
-                });
-                editor.focus();
-              }}
-            >
-              {dict.comments.save_button_text}
-            </Components.Generic.Toolbar.Button>
-          </Components.Generic.Toolbar.Root>
+        actions={({ isFocused, isEmpty }) => (
+          <FloatingComposerActionsComponent
+            isFocused={isFocused}
+            isEmpty={isEmpty}
+            onSave={onSave}
+            Components={Components}
+            dict={dict}
+          />
         )}
       />
     </Components.Comments.Card>