+++ packages/xl-multi-column/src/test/commands/moveBlocks.test.ts
@@ -28,6 +28,54 @@ describe("Test moveBlocksUp", () => {
 
     expect(getEditor().document).toMatchSnapshot();
   });
+
+  it("Selection spans column list", () => {
+    getEditor().setSelection("paragraph-1", "paragraph-2");
+
+    getEditor().moveBlocksUp();
+
+    expect(getEditor().document).toMatchSnapshot();
+  });
+
+  it("Selection across columns", () => {
+    getEditor().setSelection("column-paragraph-1", "column-paragraph-2");
+
+    getEditor().moveBlocksUp();
+
+    expect(getEditor().document).toMatchSnapshot();
+  });
+
+  it("Selection starts outside, ends in column", () => {
+    getEditor().setSelection("paragraph-1", "column-paragraph-1");
+
+    getEditor().moveBlocksUp();
+
+    expect(getEditor().document).toMatchSnapshot();
+  });
+
+  it("Selection starts outside, ends in second column", () => {
+    getEditor().setSelection("paragraph-1", "column-paragraph-2");
+
+    getEditor().moveBlocksUp();
+
+    expect(getEditor().document).toMatchSnapshot();
+  });
+
+  it("Selection starts in column, ends outside", () => {
+    getEditor().setSelection("column-paragraph-2", "paragraph-2");
+
+    getEditor().moveBlocksUp();
+
+    expect(getEditor().document).toMatchSnapshot();
+  });
+
+  it("Selection starts in first column, ends outside", () => {
+    getEditor().setSelection("column-paragraph-1", "paragraph-2");
+
+    getEditor().moveBlocksUp();
+
+    expect(getEditor().document).toMatchSnapshot();
+  });
 });
 
 describe("Test moveBlocksDown", () => {
@@ -49,6 +97,54 @@ describe("Test moveBlocksDown", () => {
 
   it("Move into next column", () => {
     getEditor().setTextCursorPosition("column-paragraph-1");
+
+    getEditor().moveBlocksDown();
+
+    expect(getEditor().document).toMatchSnapshot();
+  });
+
+  it("Selection spans column list", () => {
+    getEditor().setSelection("paragraph-1", "paragraph-2");
+
+    getEditor().moveBlocksDown();
+
+    expect(getEditor().document).toMatchSnapshot();
+  });
+
+  it("Selection across columns", () => {
+    getEditor().setSelection("column-paragraph-1", "column-paragraph-2");
+
+    getEditor().moveBlocksDown();
+
+    expect(getEditor().document).toMatchSnapshot();
+  });
+
+  it("Selection starts outside, ends in column", () => {
+    getEditor().setSelection("paragraph-1", "column-paragraph-1");
+
+    getEditor().moveBlocksDown();
+
+    expect(getEditor().document).toMatchSnapshot();
+  });
+
+  it("Selection starts outside, ends in second column", () => {
+    getEditor().setSelection("paragraph-1", "column-paragraph-2");
+
+    getEditor().moveBlocksDown();
+
+    expect(getEditor().document).toMatchSnapshot();
+  });
+
+  it("Selection starts in column, ends outside", () => {
+    getEditor().setSelection("column-paragraph-2", "paragraph-2");
+
+    getEditor().moveBlocksDown();
+
+    expect(getEditor().document).toMatchSnapshot();
+  });
+
+  it("Selection starts in first column, ends outside", () => {
+    getEditor().setSelection("column-paragraph-1", "paragraph-2");
 
     getEditor().moveBlocksDown();